refactor: move signup embed format into views dir
